pyproject.toml
[project]
name = "pia-project-energy-analysis"
version = "0.1.0"
description = "US Weather and Energy Analysis Pipeline"
authors = [
    { name = "Raphael Daveal", email = "raphaeldaveal@gmail.com" }
]
dependencies = [
    "requests",
    "pandas",
    "streamlit>=1.33.0",
    "plotly",
    "scipy",
    "python-dotenv",
    "pyyaml",
    "tenacity",
    "statsmodels",
]
requires-python = ">=3.9"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["pia_project_energy_analysis"]